[goto] fix circular dependencies in ensuring types are complete
Essentially all trash_type_symbols does, is to ensure that all types are complete
that is, a size is known.

The invariant that should hold for all types is that they are **complete**.

Previously wallop_type(name, typenames, sname) would recurse forever when
encountering a type with circular dependencies, because it never checked whether
it should actually continue the recursion!
Stopping recursion when we try to wallop_type a dependency which is the same as
`sname` (the typename at the root of the recursion) might work, but the correctness
of this approach wasn't immediately clear to me, so I discarded this.

Instead, we have to remember that the purpose of this code (even if not spelled out)
is likely to ensure that all types are complete.
So we do exactly that.
wallop_type/ensure_type_is_complete will now only recurse when its dependency is
incomplete. Because a C/C++ struct can not circularly depend on itself
(this only works via pointers and pointer types are consider complete)
termination is guaranteed.

Another solution that would likely also have worked would be to compute dependencies
more precisely. Right now, we also consider `a` to depend on `b`, if `a` has a method
that has `b` as a parameter type. But this dependency is irrelevant for whether `a`
is complete or not as `a` does not have any fields of type `b` in it.
However, the code that computes dependencies is quite general, so I chose not to change
it.