[clang] handle conversions related to boolean operators; partially revert #2344
Many boolean operators in C actually return int, because there was not bool type
at the time. The idea behind the `clang_c_convert` is to take the AST from clang
as-is and translate it to ESBMC's IR.
`clang_c_adjust_expr` then somewhat translates from the C world to the SMT world
where boolean operators only return well bools :P
To follow this design a bit more, we have to add additional casts in the adjuster.