[clang-c/cpp] move layouting a struct to adjuster
Previously we had problems with recursive structs, templates, and
forward references.
The problems existed, because we were parsing a class and then would also layout
the class. But the layout would require parsing other bases or field types
which would then cause layouting for them as well.
If we have recursive structs via pointers this would lead to infinite loops.
Various ways to solve this were devised, but always felt more like a hack.

To fix this, we now only copy the own components of a struct and
consider a struct complete as soon as clang has a definition for it.
There should be no code that relies on the layout of the struct at this point making this safe.

In the adjust phase, **all** structs have been resolved and contain their own fields.
We then just have to layout them like before.