CPROVER migration compatibility (#2443)
In this PR I am introducing ireps that were originated from CBMC into
the migration. However, I am focusing on implementing them as intrinsic
function calls. As we can then test the behavior using C files.

This is still a draft as it requires:
- [x] ~There is a bug in the transcoder that is making the pop_frame
assertion break.~ Applying same hack as coverage.
- [X] ~Overflow-result needs to invoke our intrinsic function.~ Can't
apply late goto transformations.
- [X]  ~r_ok intrinsic. ~ Will fix in another PR
- [X] .goto testcases need to be constructed from CBMC and then adapted
to ESBMC format (using goto-transcoder).
- [x] ~Export header of ESBMC. This is a C file including all the
intrinsic function signatures, this is meant to be exported as a .goto
file.~ Testing utilities will be ported to utils (#2476).

---------

Co-authored-by: Rafael Sá Menezes <rafael.sa.menezes40@gmail.com>
Co-authored-by: rafaelsamenezes <8601807+rafaelsamenezes@users.noreply.github.com>