Consider the guard when dealing with sideeffects (#2514)
This is one of the known problems:

1. The current guard needs to be taken into account when dealing with
memory allocation, especially in if expressions.
2. There is something wrong with our realloc operation that causes out
of bound. This is because we only adjust the bound in the track new
pointer, while nothing changes in the dynamic object. I tried to modify
the symbol directly, but at the SMT level there was a bit width
mismatch.

Should we rethink the implementation of realloc, perhaps malloc and
memcpy would be a good approach instead?