Recognising "global" declarations in ESBMC-Python (#2551)
The changes made allows for ESBMC-Python to recognise the "global"
keyword. 4 new tests were written to ensure that ESBMC-Python produces
the correct verification results/errors as Python does.

Notably, all symbols generated had their static_lifetime flag set to
false regardless of scope prior to this pull request. I had changed it
to only be false if the variable is within a class or a function.

---------

Signed-off-by: Tew Yi Jun <tewyijun@gmail.com>
Co-authored-by: Lucas C. Cordeiro <lucasccordeiro@gmail.com>