Updated memcpy function (#2561)
Improved the memcpy function to store 8 bytes per iteration and check
8-byte alignment for pointers
If the pointers are aligned, the function safely, it copies 8 bytes of
memory using uint64_t pointers for efficiency. Remaining bytes are
handled in a trailing byte loop.
If the pointers are not 8-bytes aligned, the function falls back to a
safe byte-by-byte copy.
Addressed issue https://github.com/esbmc/esbmc/issues/2282
This is the continuation of #2545

---------

Signed-off-by: mihaistate05 <mihai.state03@yahoo.com>
Co-authored-by: intrigus-lgtm <60750685+intrigus-lgtm@users.noreply.github.com>
Co-authored-by: XLiZHI <958689657@qq.com>
Co-authored-by: Lucas C. Cordeiro <lucasccordeiro@gmail.com>