[python] add support for input() built-in function
Fixes crash when Python programs use input() function. Previously,
input() was not recognized as a built-in function and fell through
to symbol lookup, causing "migrate expr failed" error and abort.

This PR:
- Adds is_input_call() method to detect input() function calls
- Adds handle_input() method to model input() as non-deterministic string
- Insert input() handling in get() method before built-in type checks

The input() function is now modeled as returning a non-deterministic
string with maximum length of 256 characters, which is appropriate
an under-approximation.

Resolves: ESBMC crash on programs using input() function

Signed-off-by: Lucas Cordeiro <lucasccordeiro@gmail.com>