with expression constant propagation (Experiment) (#2412)
This PR (30s):

```
Statistics:          33569 Files
  correct:           17188
    correct true:    10029
    correct false:    7159
  incorrect:            36
    incorrect true:     14
    incorrect false:    22
  unknown:           16345
  Score:             26417 (max: 55885)

https://github.com/esbmc/esbmc/actions/runs/14686517130/job/41216206087
```

Master (30s):
```
Statistics:          33569 Files
  correct:           17623
    correct true:    10345
    correct false:    7278
  incorrect:            35
    incorrect true:     14
    incorrect false:    21
  unknown:           15911
  Score:             27184 (max: 55885)

https://github.com/esbmc/esbmc/actions/runs/14686515266/job/41216206606
```

Causes some performance impact in some benchmarks, which I summarize as
the following example:
```
#include <stdlib.h>

int main()
{
  int *dataBuffer = malloc(400UL);
  size_t i;

  i = 0;

  goto ldv_3210;

ldv_3209:
  *(dataBuffer + i) = 5;

  i += 1;

ldv_3210:
  if (i <= 99)
    goto ldv_3209;

  free(dataBuffer);

  return 0;
}
```
In this example, as the loop executes, the symbol `dataBuffer` ends up
forming a giant with chain:
```
with 
     with
             with
                     ....
```
When we don't use constant propagation the backend doesn't have to deal
with this huge with chain. In addition, we use incremental bmc in
SV-COMP, which makes us deal with it repeatedly, which leads to
problems.

Whether we have a way to get rid of the with chain, constant propagation
does not seem to play any role in the above example.

---------

Signed-off-by: Lucas Cordeiro <lucasccordeiro@gmail.com>
Co-authored-by: Lucas Cordeiro <lucasccordeiro@gmail.com>
Co-authored-by: lucasccordeiro <3694109+lucasccordeiro@users.noreply.github.com>