Memcpy optimization (#2894)
This pull request introduces an intrinsic implementation for the
`memcpy` function in the symbolic execution engine, enabling more
precise modeling and verification of memory copy operations in C
programs. Currently it only works for primitives, and it has an extra
preprocessing cost (which makes it slower for small memcpy and faster
for big memcpy).

30s run:
```
Statistics:          33569 Files
  correct:           17695
    correct true:    10385
    correct false:    7310
  incorrect:            22
    incorrect true:     11
    incorrect false:    11
  unknown:           15852
  Score:             27552 (max: 55885)
```

This is a draft as I am doing another 30s run with this branch and
master just to be extra sure.

---------

Co-authored-by: Rafael Sá Menezes <rafael.sa.menezes40@gmail.com>
Co-authored-by: rafaelsamenezes <8601807+rafaelsamenezes@users.noreply.github.com>