add `--dont-care-about-missing-extensions`
This option is designed to make ESBMC continue processing even if it
encounters unsupported extensions during the compilation process,
rather than crashing.

This could be useful in scenarios where handling these unsupported
extensions is not critical and it's more important to continue with
the verification process, even if some aspects of the code might not
be correctly verified due to missing support for certain extensions.