[python] fix ternary expression array-to-pointer conversion for pointer assignments (#3039)
Fixes https://github.com/esbmc/esbmc/issues/3038.

This PR converts each branch to a pointer before building the
`if`-expression when a ternary expression with array operands (e.g.,
string literals) is assigned to a pointer member. This prevents
`compute_pointer_offset` errors when the backend tries to take the
address of an if expression with an array type.

Thanks to [zhassan-aws](https://github.com/zhassan-aws) for reporting
this issue.