[python] resolve forward reference for method return types in classes (#3160)
Fixes https://github.com/esbmc/esbmc/issues/3155.

This PR looks up the return type from the AST when the called function's
symbol is not yet in the symbol table, ensuring the correct type is used
for temporary variables even with forward references.

When a method calls another method defined later in the class, the
return type is unknown during conversion, causing a temporary variable
to be created with void type instead of the correct return type; this
led to a crash during symbolic execution.

Thanks to [zhassan-aws](https://github.com/zhassan-aws) for reporting
this issue.

Signed-off-by: Lucas Cordeiro <lucasccordeiro@gmail.com>